refactor:点赞返回积分，修改绑定银行卡接口

+++ service-share/ad-srv/src/main/java/com/lawu/eshop/ad/srv/service/AdService.java
@@ -1,5 +1,7 @@
 package com.lawu.eshop.ad.srv.service;
 
+import java.math.BigDecimal;
+
 import com.lawu.eshop.ad.param.AdMemberParam;
 import com.lawu.eshop.ad.param.AdMerchantParam;
 import com.lawu.eshop.ad.param.AdPraiseParam;
@@ -87,7 +89,7 @@ public interface AdService {
 	 * @param memberId
 	 * @return
 	 */
-	Integer clickPraise(Long id,Long memberId,String num);
+	BigDecimal clickPraise(Long id,Long memberId,String num);
 	
 	/**
 	 * 点击广告
